+++ tests/test_sslaestransport.py
@@ -11,7 +11,6 @@ import aiohttp
 import pytest
 from yarl import URL
 
-from kasa.aestransport import AesEncyptionSession
 from kasa.credentials import Credentials
 from kasa.deviceconfig import DeviceConfig
 from kasa.exceptions import (
@@ -26,6 +25,7 @@ from kasa.experimental.sslaestransport import (
 )
 from kasa.httpclient import HttpClient
 from kasa.protocol import DEFAULT_CREDENTIALS, get_default_credentials
+from kasa.transports.aestransport import AesEncyptionSession
 
 # Transport tests are not designed for real devices
 pytestmark = [pytest.mark.requires_dummy]
